backend/requirements.txt
Flask==2.1.1
SQLAlchemy==1.4.27
pydantic==2.5.2
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==0.19.2
alembic==1.7.7
requests==2.26.0
passlib==1.7.4
bcrypt==3.2.0
python-jose==3.3.0
jinja2>=3.0.0
python-multipart>=0.0.5
email-validator>=1.1.3
httpx>=0.23.0